сделали красиво всё в БД
