+++ helpers.php
@@ -126,4 +126,21 @@ function get_noun_plural_form (int $number, string $one, string $two, string $ma
  * @param array $data Ассоциативный массив с данными для шаблона
  * @return string Итоговый HTML
  */
+function include_template($name, array $data = []) 
+{
+    $name = 'templates/' . $name;
+    $result = '';
 
+    if (!is_readable($name)) 
+    {
+        return $result;
+    }
+
+    ob_start();
+    extract($data);
+    require $name;
+
+    $result = ob_get_clean();
+
+    return $result;
+}
