+++ system/init.php
@@ -1,5 +1,8 @@
 <?php
-require_once('function/function.php');
+ini_set('error_reporting', E_ALL); 
+ini_set('display_errors', 1); 
+ini_set('display_startup_errors', 1);
+require_once('functions/function.php');
 
 date_default_timezone_set('Europe/Moscow');
 $title = "Дела в порядке";
